Move iou to its own file and update tests to match
